+++ TYPESCRIPT_REWRITE_PLAN.md
@@ -1,0 +1,88 @@
+# KONIVRER Deck Database - Complete TypeScript Rewrite
+
+## Overview
+This document outlines the comprehensive TypeScript rewrite of the KONIVRER Deck Database repository to eliminate all 12,797 TypeScript errors and modernize the codebase.
+
+## Current Issues
+- 12,797 TypeScript errors across 166 files
+- JSX syntax errors (missing parent elements, unclosed tags)
+- Invalid TypeScript syntax (malformed functions, incomplete expressions)
+- Missing type definitions
+- Deprecated code patterns
+
+## Rewrite Strategy
+
+### Phase 1: Core Infrastructure (Priority 1)
+1. **Type Definitions** (`src/types/`)
+   - Consolidate and modernize all type definitions
+   - Create comprehensive interfaces for all data structures
+   - Add strict typing for all API responses
+
+2. **Core Components** (`src/components/`)
+   - Fix JSX syntax errors
+   - Add proper TypeScript interfaces
+   - Modernize component patterns (hooks, functional components)
+
+3. **Context Providers** (`src/contexts/`)
+   - Fix all context implementations
+   - Add proper TypeScript generics
+   - Ensure type safety across providers
+
+### Phase 2: Pages and Features (Priority 2)
+1. **Page Components** (`src/pages/`)
+   - Rewrite all page components with proper TypeScript
+   - Fix JSX structure issues
+   - Add proper error boundaries
+
+2. **Services** (`src/services/`)
+   - Modernize all service implementations
+   - Add proper async/await patterns
+   - Implement comprehensive error handling
+
+### Phase 3: Game Engine and Advanced Features (Priority 3)
+1. **Game Engine** (`src/engine/`)
+   - Rewrite game logic with strict typing
+   - Modernize AI implementations
+   - Add comprehensive game state management
+
+2. **Utilities and Hooks** (`src/utils/`, `src/hooks/`)
+   - Modernize all utility functions
+   - Add proper TypeScript generics
+   - Implement comprehensive testing
+
+## Implementation Plan
+
+### Step 1: Clean Slate Approach
+- Remove all broken components temporarily
+- Create new, properly typed implementations
+- Ensure each component compiles before moving to next
+
+### Step 2: Modern TypeScript Patterns
+- Use latest TypeScript 5.4+ features
+- Implement strict null checks
+- Use proper generic constraints
+- Add comprehensive JSDoc comments
+
+### Step 3: Code Quality
+- Implement ESLint strict rules
+- Add Prettier formatting
+- Ensure 100% TypeScript compliance
+- Add comprehensive testing
+
+## Success Criteria
+- Zero TypeScript errors
+- All components properly typed
+- Modern React patterns throughout
+- Comprehensive test coverage
+- Clean, maintainable codebase
+
+## Timeline
+- Phase 1: Core Infrastructure (Immediate)
+- Phase 2: Pages and Features (Next)
+- Phase 3: Advanced Features (Final)
+
+## Quality Assurance
+- Continuous TypeScript checking
+- Automated testing
+- Code review process
+- Performance monitoring
