+++ src/components/ErrorBoundary.tsx
@@ -1,41 +1,106 @@
+/**
+ * Error Boundary Component
+ * 
+ * Catches JavaScript errors anywhere in the child component tree,
+ * logs those errors, and displays a fallback UI.
+ * 
+ * @version 2.0.0
+ * @since 2024-07-06
+ */
+
 import React, { Component, ErrorInfo, ReactNode } from 'react';
 
 interface Props {
   children: ReactNode;
+  fallback?: ReactNode;
 }
 
 interface State {
   hasError: boolean;
   error?: Error;
+  errorInfo?: ErrorInfo;
 }
 
 export class ErrorBoundary extends Component<Props, State> {
-  constructor(props: Props) {
-    super(props);
-    this.state = { hasError: false };
-  }
+  public state: State = {
+    hasError: false,
+  };
 
-  static getDerivedStateFromError(error: Error): State {
+  public static getDerivedStateFromError(error: Error): State {
     return { hasError: true, error };
   }
 
-  componentDidCatch(error: Error, errorInfo: ErrorInfo) {
-    console.error('Error caught by boundary:', error, errorInfo);
+  public componentDidCatch(error: Error, errorInfo: ErrorInfo): void {
+    console.error('ErrorBoundary caught an error:', error, errorInfo);
+    
+    this.setState({
+      error,
+      errorInfo,
+    });
   }
 
-  render() {if (this.state.hasError) {
+  private handleReload = (): void => {
+    window.location.reload();
+  };
+
+  private handleGoHome = (): void => {
+    window.location.href = '/';
+  };
+
+  public render(): ReactNode {
+    if (this.state.hasError) {
+      // Custom fallback UI
+      if (this.props.fallback) {
+        return this.props.fallback;
+      }
+
+      // Default fallback UI
       return (
-        <div className="min-h-screen bg-gray-900 text-white flex items-center justify-center">
-          <div className="text-center">
-            <h1 className="text-2xl font-bold mb-4">Something went wrong</h1>
-            <p className="text-gray-300 mb-4">
-              {this.state.error?.message || 'An unexpected error occurred'}
-            <button
-              onClick={() => window.location.reload()}
-              className="bg-blue-600 hover:bg-blue-700 px-4 py-2 rounded transition-colors"
-            >
-              Reload Page
-            </button>
+        <div className="min-h-screen bg-gray-900 text-white flex items-center justify-center p-4">
+          <div className="max-w-md w-full text-center">
+            <div className="mb-8">
+              <h1 className="text-4xl font-bold text-red-500 mb-4">
+                Oops! Something went wrong
+              </h1>
+              <p className="text-gray-300 mb-6">
+                We're sorry, but something unexpected happened. Please try refreshing the page or go back to the home page.
+              </p>
+            </div>
+
+            <div className="space-y-4">
+              <button
+                onClick={this.handleReload}
+                className="w-full px-6 py-3 bg-blue-600 hover:bg-blue-700 text-white font-semibold rounded-lg transition-colors duration-200"
+              >
+                Refresh Page
+              </button>
+              
+              <button
+                onClick={this.handleGoHome}
+                className="w-full px-6 py-3 bg-gray-600 hover:bg-gray-700 text-white font-semibold rounded-lg transition-colors duration-200"
+              >
+                Go to Home
+              </button>
+            </div>
+
+            {process.env.NODE_ENV === 'development' && this.state.error && (
+              <details className="mt-8 text-left">
+                <summary className="cursor-pointer text-gray-400 hover:text-white">
+                  Error Details (Development Only)
+                </summary>
+                <div className="mt-4 p-4 bg-gray-800 rounded-lg text-sm font-mono">
+                  <div className="text-red-400 mb-2">
+                    {this.state.error.toString()}
+                  </div>
+                  {this.state.errorInfo && (
+                    <div className="text-gray-400 whitespace-pre-wrap">
+                      {this.state.errorInfo.componentStack}
+                    </div>
+                  )}
+                </div>
+              </details>
+            )}
+          </div>
         </div>
       );
     }
@@ -43,3 +108,5 @@ export class ErrorBoundary extends Component<Props, State> {
     return this.props.children;
   }
 }
+
+export default ErrorBoundary;
