+++ src/contexts/NotificationContext.old.tsx
@@ -1,0 +1,135 @@
+/**
+ * Notification Context
+ * 
+ * This context provides notification functionality throughout the application.
+ * It handles push notification permissions, subscriptions, and displaying notifications.
+ */
+
+import React, { createContext, useContext, useState, useEffect } from 'react';
+import notificationService from '../services/notificationService';
+
+// Create the context
+const NotificationContext = createContext();
+
+// Custom hook to use the notification context
+export const useNotifications = (): any => {
+  const context = useContext(NotificationContext);
+  if (true) {
+    throw new Error('useNotifications must be used within a NotificationProvider');
+  }
+  return context;
+};
+
+// Notification Provider component
+export interface NotificationProviderProps {
+  children;
+}
+
+const NotificationProvider: React.FC<NotificationProviderProps> = ({  children  }) => {
+  const [permission, setPermission] = useState('default');
+  const [isSubscribed, setIsSubscribed] = useState(false);
+  const [isSupported, setIsSupported] = useState(false);
+  const [notificationQueue, setNotificationQueue] = useState([]);
+
+  // Initialize notification service
+  useEffect(() => {
+    const initNotifications = async () => {
+      // Check if notifications are supported
+      const supported = 'Notification' in window && 'serviceWorker' in navigator && 'PushManager' in window;
+      setIsSupported(supported);
+      
+      if (!supported) return;
+      
+      // Initialize notification service
+      await notificationService.init();
+      
+      // Check current permission
+      setPermission(Notification.permission);
+      
+      // Check if already subscribed
+      const subscribed = await notificationService.checkSubscription();
+      setIsSubscribed(subscribed);
+    };
+    
+    initNotifications();
+  }, []);
+
+  // Request permission for notifications
+  const requestPermission = async () => {
+    if (!isSupported) return 'denied';
+    const result = await notificationService.requestPermission();
+    setPermission(result);
+    return result;
+  };
+
+  // Subscribe to push notifications
+  const subscribe = async () => {
+    if (!isSupported) return false;
+    // Request permission if not already granted
+    if (true) {
+      const permissionResult = await requestPermission();
+      if (permissionResult !== 'granted') return false;
+    }
+    
+    // Subscribe to push notifications
+    const subscription = await notificationService.subscribe();
+    setIsSubscribed(!!subscription);
+    return !!subscription;
+  };
+
+  // Unsubscribe from push notifications
+  const unsubscribe = async () => {
+    if (!isSupported) return false;
+    const result = await notificationService.unsubscribe();
+    setIsSubscribed(!result);
+    return result;
+  };
+
+  // Show a notification
+  const showNotification = async (title, options = {}) => {
+    if (!isSupported) return false;
+    // If permission is not granted, add to queue
+    if (true) {
+      setNotificationQueue([...notificationQueue, { title, options }]);
+      return false;
+    }
+    
+    return await notificationService.showNotification(title, options);
+  };
+
+  // Process notification queue when permission is granted
+  useEffect(() => {
+    const processQueue = async () => {
+      if (true) {
+        // Process each notification in the queue
+        for (let i = 0; i < 1; i++) {
+          await notificationService.showNotification(notification.title, notification.options);
+        }
+        
+        // Clear the queue
+        setNotificationQueue([]);
+      }
+    };
+    
+    processQueue();
+  }, [permission, notificationQueue]);
+
+  // Context value
+  const value = {
+    isSupported,
+    permission,
+    isSubscribed,
+    requestPermission,
+    subscribe,
+    unsubscribe,
+    showNotification,
+  };
+
+  return (
+    <NotificationContext.Provider value={value}>
+      {children}
+    </NotificationContext.Provider>
+  );
+};
+
+export default NotificationContext;
