+++ src/styles/golden-menu.css
@@ -1,0 +1,151 @@
+/**
+ * Golden Menu Bar Styles
+ * 
+ * Styles for the horizontal navigation bar with golden/bronze theme
+ */
+
+.golden-menu-bar {
+  background: linear-gradient(135deg, #2a2a2a 0%, #1a1a1a 100%);
+  border-bottom: 2px solid #3a3a3a;
+  padding: 0;
+  position: relative;
+  overflow: hidden;
+}
+
+.golden-menu-bar::before {
+  content: '';
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background: linear-gradient(90deg, 
+    rgba(184, 134, 11, 0.1) 0%, 
+    rgba(217, 119, 6, 0.1) 25%, 
+    rgba(245, 158, 11, 0.1) 50%, 
+    rgba(217, 119, 6, 0.1) 75%, 
+    rgba(184, 134, 11, 0.1) 100%
+  );
+  pointer-events: none;
+}
+
+.golden-menu-container {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  max-width: 1200px;
+  margin: 0 auto;
+  padding: 0 2rem;
+  position: relative;
+  z-index: 1;
+}
+
+.golden-menu-items {
+  display: flex;
+  align-items: center;
+  gap: 0;
+}
+
+.golden-menu-item {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 1rem 2rem;
+  text-decoration: none;
+  color: #d4af37;
+  transition: all 0.3s ease;
+  position: relative;
+  cursor: pointer;
+  border-radius: 8px;
+  margin: 0 0.25rem;
+}
+
+.golden-menu-item:hover {
+  color: #f4d03f;
+  background: rgba(212, 175, 55, 0.1);
+  transform: translateY(-2px);
+}
+
+.golden-menu-item.active {
+  color: #f4d03f;
+  background: rgba(212, 175, 55, 0.15);
+}
+
+.golden-menu-item.active::after {
+  content: '';
+  position: absolute;
+  bottom: 0;
+  left: 50%;
+  transform: translateX(-50%);
+  width: 60%;
+  height: 3px;
+  background: linear-gradient(90deg, #d4af37, #f4d03f, #d4af37);
+  border-radius: 2px;
+}
+
+.golden-menu-icon {
+  margin-bottom: 0.5rem;
+  filter: drop-shadow(0 0 4px rgba(212, 175, 55, 0.3));
+}
+
+.golden-menu-label {
+  font-size: 0.875rem;
+  font-weight: 500;
+  letter-spacing: 0.5px;
+  text-transform: uppercase;
+  text-shadow: 0 0 8px rgba(212, 175, 55, 0.3);
+}
+
+.golden-menu-auth {
+  display: flex;
+  align-items: center;
+}
+
+.login-item {
+  background: rgba(212, 175, 55, 0.1);
+  border: 1px solid rgba(212, 175, 55, 0.3);
+  border-radius: 8px;
+}
+
+.login-item:hover {
+  background: rgba(212, 175, 55, 0.2);
+  border-color: rgba(212, 175, 55, 0.5);
+}
+
+.golden-menu-profile {
+  position: relative;
+}
+
+/* Responsive design */
+@media (max-width: 768px) {
+  .golden-menu-container {
+    padding: 0 1rem;
+  }
+  
+  .golden-menu-item {
+    padding: 0.75rem 1rem;
+  }
+  
+  .golden-menu-label {
+    font-size: 0.75rem;
+  }
+  
+  .golden-menu-icon {
+    margin-bottom: 0.25rem;
+  }
+}
+
+@media (max-width: 640px) {
+  .golden-menu-items {
+    gap: 0;
+  }
+  
+  .golden-menu-item {
+    padding: 0.5rem 0.75rem;
+    margin: 0;
+  }
+  
+  .golden-menu-label {
+    font-size: 0.7rem;
+  }
+}
